Compile with clang when C++11 support is enabled
text/qtextdocument_p.cpp:1070:43: error: non-constant-expression cannot
be narrowed from type 'int' to 'quint32' (aka 'unsigned int') in
initializer list [-Wc++11-narrowing]
text/qtextdocument_p.cpp:1070:43: note: override this message by
inserting an explicit cast


cregit: 6deebc817080f9eb86d5d6fbe50444678eed4dfe https://github.com/qt/qtbase/commit/6deebc817080f9eb86d5d6fbe50444678eed4dfe

+++ src/gui/text/qtextdocument_p.cpp
@@ -111,7 +111,7 @@ parameter_list|,
 name|a8
 parameter_list|)
 define|\
-value|QTextUndoCommand c = { a1, a2, 0, 0, quint8(a3), a4, a5, a6, { a7 }, a8 }
+value|QTextUndoCommand c = { a1, a2, 0, 0, quint8(a3), a4, quint32(a5), quint32(a6), { int(a7) }, quint32(a8) }
 else|#
 directive|else
 define|#
